requirements.txt
agentic-doc==0.0.16
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
Django==4.2.20
exceptiongroup==1.2.2
fpdf==1.7.2
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
narwhals==1.32.0
numpy==1.24.4
openai==1.70.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pillow_heif==0.16.0
pkgutil_resolve_name==1.3.10
playwright==1.48.0
propcache==0.2.0
protobuf==5.29.4
pyarrow==17.0.0
pydantic==2.7.4
pydantic-settings==2.9.1
pydantic_core==2.18.4
pydeck==0.9.1
pyee==12.0.0
Pygments==2.19.1
PyMuPDF==1.25.5
pypdf==5.4.0
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
selenium==4.27.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
streamlit==1.40.1
structlog==25.2.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.2.3
watchdog==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.15.2
zipp==3.20.2
